# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Shorten Url API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/decode": {
            "post": {
                "description": "API to decodes a shortened URL to its original URL",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ShortLink"
                ],
                "summary": "Decodes a shortened URL to its original URL",
                "parameters": [
                    {
                        "description": "Request payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ConvertUrlRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.DecodeUrlResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputils.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputils.Response"
                        }
                    }
                }
            }
        },
        "/encode": {
            "post": {
                "description": "API to encodes a url to a shorten url",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ShortLink"
                ],
                "summary": "Encodes a url to a shorten url",
                "parameters": [
                    {
                        "description": "Request payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ConvertUrlRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.EncodeUrlResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputils.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputils.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.ConvertUrlRequest": {
            "type": "object",
            "properties": {
                "url": {
                    "type": "string"
                }
            }
        },
        "dto.DecodeUrlResponse": {
            "type": "object",
            "properties": {
                "originalUrl": {
                    "type": "string"
                }
            }
        },
        "dto.EncodeUrlResponse": {
            "type": "object",
            "properties": {
                "shortenUrl": {
                    "type": "string"
                }
            }
        },
        "httputils.Response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    }
}